Started verification optimization

+++ Master/ValidationFormulaire/ValidationFormulaire.Core/Interfaces/ImageOperations/IComparePixelColor.vb
@@ -1,5 +1,5 @@
 ﻿Public Interface IComparePixelColor
-    Function DifferentPixelColor(color1a As Integer, color2a As Integer, color1b As Integer,
-                                        color2b As Integer, color1c As Integer, color2c As Integer,
+    Function DifferentPixelColor(color1a As Byte, color2a As Byte, color1b As Byte,
+                                        color2b As Byte, color1c As Byte, color2c As Byte,
                                         Optional color_threshold As Integer = 50) As Boolean
 End Interface
